+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/application/TeacherRestController.java
@@ -38,6 +38,11 @@ public class TeacherRestController {
         return teacherController.getTeacherCourses(teacherId);
     }
 
+    @GetMapping("")
+    public List<Teacher> getAllTeachers(){
+        return teacherController.getAllTeachers();
+    }
+
     @PutMapping("/{teacherId}/addCourse/{courseId}")
     public Teacher addCourseToTeacher(@PathVariable Long teacherId, @PathVariable Long courseId){
         return teacherController.addCourseToTeacher(teacherId, courseId);