+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/SchoolController.java
@@ -4,6 +4,7 @@ import com.nimio2024.nimio2024_tfg_polmorenocastelar.application.dto.SchoolDTO;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Center;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Course;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.School;
+import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Teacher;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.service.CenterService;
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.service.SchoolService;
 import org.springframework.stereotype.Controller;
@@ -69,5 +70,10 @@ public class SchoolController {
         School school = schoolService.getSchoolById(schoolId);
         school.setSchoolName(schoolName);
         return schoolService.saveSchool(school);
+    }
+
+    public List<Teacher> getTeachersFromSchool(Long schoolId) {
+        School school = schoolService.getSchoolById(schoolId);
+        return school.getTeachers();
     }
 }