+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/controllers/TeacherController.java
@@ -78,4 +78,8 @@ public class TeacherController {
 
 
     }
+
+    public List<Teacher> getAllTeachers() {
+        return teacherService.getAllTeachers();
+    }
 }