+++ NIMIO2024_TFG_PolMorenoCastelar/src/main/java/com/nimio2024/nimio2024_tfg_polmorenocastelar/persistence/TeacherRepository.java
@@ -2,8 +2,13 @@ package com.nimio2024.nimio2024_tfg_polmorenocastelar.persistence;
 
 import com.nimio2024.nimio2024_tfg_polmorenocastelar.domain.Teacher;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 @Repository
 public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+
+
 }